requirements.txt
streamlit
openai
requests
python-dotenv
pandas
splunk-sdk
